+++ src/XUnity.AutoTranslator.Plugin.Core/Constants/KnownEvents.cs
@@ -10,5 +10,6 @@ namespace XUnity.AutoTranslator.Plugin.Core.Constants
       public static string OnUnableToTranslateUGUI = "OnUnableToTranslateUGUI";
       public static string OnUnableToTranslateTextMeshPro = "OnUnableToTranslateTextMeshPro";
       public static string OnUnableToTranslateNGUI = "OnUnableToTranslateNGUI";
+      public static string OnUnableToTranslateIMGUI = "OnUnableToTranslateIMGUI";
    }
 }